+++ packages/applications/calculator/src/i18n/translations/en.json
@@ -22,6 +22,7 @@
   "toolbar.resetAll": "Reset All",
   "toolbar.definitions": "Definitions",
   "toolbar.history": "History",
+  "toolbar.functions": "Other functions",
 
   "settings.interface": "Interface",
   "settings.appearance": "Appearance",
